Deep links: fix creation of links to layouts w/ stack-groups
Stack groups are used to represent related stacks in order to allow the
front-end to be intelligent about their display, e.g. ortho-views or
multi-channel images. Creating deep links to layouts where a stack group
was loaded failed, because of a typo. This is fixed now.

Thanks to @acardona for reporting this issue!

Fixes catmaid/catmaid#2215