Make batch iterator PEP 479 compliant
In Python 3.7 a small Python behavior change [1] lead to runtime errors
when `StopIteration` was raised within a generator. This is described in
more detail in PEP 479 [2]. For CATMAID this caused the grid cache
update function to fail. As suggested in PEP 479, this has been fixed by
not raising `StopIteration`, but simply returning from the generator.

[1] https://docs.python.org/3/whatsnew/3.7.html#changes-in-python-behavior
[2] https://peps.python.org/pep-0479/