Create temp directory for matplotlib and opt. delete it on exit
Otherwise, matplotlib will create a temp dir in /tmp for each process,
which can end up being a lot of temp directories. Since these folders
seem to be able to be used in a shared environment (multiple processes),
the same folder is used for all processes.

Optionally, if `REMOVE_MPLCONFIGDIR_ON_EXIT` is set to `True`, the
folder can be removed on exit.