Annotations: show target annotation name(s) when adding annotation
The "New Annotation" dialog now also shows all target neurons. This is
usually one, but can be multiple (e.g. when opened through Neuron
Navigator). To allow double checking the correct target neuron, the name
is now displayed in the lower half of the widget.

If another object is annotated than a neuron, this meta data line will
not be shown.