New widget "Data Plot": import a 3-column CSV file with skeleton_id, X value, Y value and obtain an interactive SVG plot like that of the CircuitGraphPlot.
Will filter skeleton IDs for those that actually exist, and warn you
in the javascript console about those that don't.

The intended use case is the interactive analysis of data crunched elsewhere,
for example in a python script. The widget is a SkeletonSource and
therefore, once selected, skeletons can be appended to other widgets.

For convenience, there is a "Select all" button.