SWC import: fix bug with wrong default type for parent ID = NULL
On some systems it was observed that the import of some SWC files fail.
It turns out Postgres was assuming the type ::text for the NULL parent ID
in case of the root node being stored first. To ensure the first parent
ID element is properly typed, the ::bigint type case is now added to the
first parent ID provided. This allows Postgres to use a more reasonable
default for all other values. This is faster than adding a type case to
every single parent ID element.