Celery: be more fault-tolerant on startup
Celery needs an async messaging broker like RabbitMQ or Redis. If this
can't be found or is configured in an unexpected way, an error is
raised. This could prevent the rest of CATMAID to start up. Instead, the
error is now logged and startup continued.

Also, the default expectation for the exchange configuration has been
changed to non-durable. This seems to be much more common.