Annotation retrieval: fix modified-since condition to include removal
Some front-end services like the Neuron Name Service keep track when
they last updated (meta-)annotations and on a refresh() they ask the
back-end to only return results when the annotations changed since the
last update. The back-end SQL implementation of this wasn't working
correctly: it didn't take into account the removal of annotations and
meta-annotations. This is fixed now by also looking at changes in the
past using the history tables. With this in place, the modified-since
check works as expected and e.g. the Neuron Name Service in the
front-end will refresh names correctly when an annotation is removed.

Fixes catmaid/catmaid#2237