Stack viewer: add layer blending window
The layer controls now provide an option to configure a blending window
for PIXI layers. A set of layers that reflect the sections surrounding
the current layer is blended together. A blending function can be
defined. This allows for instance minimum intensity projection of the
next 10 layers.

Please enter the commit message for your changes. Lines starting