Graph widget: add an option for mouse wheel sensitivity
This adds a new "View" tab to the Graph Widget, which is meant for
settings concerning the graph display and the interaction with the
graph. For now it only contains a new setting for mouse wheel
sensitivity. This setting is now lowered by default, because many
different users reported zooming would be too fast using the mouse
wheel.