Landmark widget: add re-use button for each stored transform
Each transformation in the list of active display transformation comes
now with a "Re-use" button in the last column. This pre-populates the
controls for a new transformation according to the transformation of the
respective table row. The following fields are updated: source remote,
source project, source group(s), target group(s), transform model and
source annotation (if any). Also the UI mode (same project, other
project, remote CATMAID) are changed according to the re-used display
transformation.