Review widget: improve handling of node deletions
When a node is deleted in a skeleton under review, the Review Widget
will now refresh the segment respresentation and move on to the next
possible node in the segment.

Also, a warning is now shown if a node can't be found on the server when
marking it reviewed, assuming it might be deleted. This, however, should
usually not be visible, because the segments should have up-to-date
information.

Fixes catmaid/catmaid#2270