Reduce number of repeated redraws on initial load (esp. with deep links)
A number of tile layer redraws have been issued initially, especially
when loading deep links (e.g. one on initial unzoomed layer-add, one for
zoom and going to right location, one after resizing due to layouting,
...). A stack layer now has an additional property named 'hidden'. When
true, the layer won't be rendered, even with visible = false. Initially
this is set to true and only once all initialization tasks are done, the
layer is shown again and a redraw is issued.