Box selection tool: make pixel conversion consistent with back-end
Up to now it could happen that a user selects a bounding box in the
front-end with a certain pixel size, that, when used for cropping,
produced an output image with off-by-one dimensions. Doing the
conversation in the same order the back-end does, fixes this.

Also slightly change the format of the returned object.