Cropping tool: fix occasional black borders in export at right/bottom
The image parts from all tiles/blocks are placed on the target canvas by
the cropping tool. The locations of each image parts were slightly off,
because the respective min/max locations in each tile was computed not
as zero-based index but as one-based index. Additionally, the
destination location of each part in the target canvas was accidentally
computed wrongly, because it targeted the last pixel of the previous
part as start of the new part, rather than targeting the next pixel to
avoid an overlap. Both is fixed now.