Only include Sphinx rather than all doc requirements in the main file
The doc requirements are mainly used for readthedocs.org, where we don't
want all our dependencies installed, because it takes too much time.
Therefore, we can't remove it. Including it using a -r reference however
breaks GitHub actions (due to extrareq it seems). To support some basic
HTML API documentation for some endpoints, Sphinx is copied directly
into the main file, because it is the only thing needed for the HTML API
docs on certain endpoints, it seems.

The downside is that the Sphinx version has to be updated in both files,
requirements.txt and requirement-doc.txt.