Review widget: fix error in node deletion handling
After node deletion, the review segment showed an error if a subarbor
was under review. This is fixed now by checking if there is a subarbor
under review in the first place, before attempting to refer to it.

Second, the view isn't moved anymore after a node was deleted. As
@acardona pointed out, this is rather distracting and one is usually
well aware where one deletes a node. Plus, returning to the node under
review can be achieved by W/E key presses or by clicking on the "Review"
button.

See catmaid/catmaid@2280