Init: make loading layouts more robust
If a deep link was loaded without being logged in, an error could be
raised during layout construction, because referenced stack viewers are
not available. This is fixed now and a simple warning is shown.

Also the layout loading error has been simplified and doesn't include
the complete layout spec anymore. The spec didn't contribute to the
readability.