Stack layer: don't misuse visibility parameter for property "hidden"
A hidden layer is pretended to be not there. Invisible layers are
attempted to render and tiles are rendered with an opacity of zero. When
constructing a new stack layer which is asked to be hidden, now only the
field "hidden" is set. This makes it possible again to load and reload
stacks from deep links and after resizes.