Stack viewer: fix reset of flashing reference lines
After flashing reference lines their visibility is reset to the state
before flashing. When storing a reference to this state by reading the
present visibility, it can happen that the wrong value is stored because
of rendering delays in the browser. This leads to reference lines
staying visible even though they are disabled in the session settings.
To prevent this from happening a reference is now stored from the actual
widget settings.