Landmark widget: allow clicks on target locations in transform table
The target of a transformation is always the current CATMAID project.
When clicking on a target location in the point matches column of the
transform table, the view will now jump to the respective location. This
is currently only implemented for the target location, not the source.

See catmaid/catmaid@2287