Landmark widget: add option to export of transform as JS function
Each created transform can now be exported as a JavaScript function by
clicking the "Export JS" button. This will create a transform() function
that takes a [x, y, z] array and returns a transformed point [x', y',
z']. The function is provided as a JavaScript file, is printed to
console and copied to clipboard. This is useful for debugging
transformations.

See catmaid/catmaid#2287