Landmark widget: add dialog to edit used landmarks per created transform
The list of actually used landmarks of the set of matching landmarks in
a transform can be edited. By clicking the "edit" button next to the
list of name matches of a transform, a dialog allows to select the
landmarks that ere used. The transform table reflects changes made here
and so do the 3D Viewer and the Landmark Layer. This makes identifying
problematic landmarks much quicker.

See catmaid/catmaid#2287