Log/history widget: add button to list skeletons in transaction
Each entry in the list of transactions in the History tab of the
Log/history widget has now a button to show all (historic) skeleton IDs
involved with this transaction. Depending on the transaction label this
can be different things (e.g. all linked skeletons in a connector
removal).

See catmaid/catmaid@2291