Image blocks: make lading sets of image blocks more robust
When image blocks are loaded, the resulting list of image blocks might
be smaller than the original list, due to merging requests. This can
also lead to the fact, that the list is completely empty, if it turns
out there is nothing to load. In case a deduplication requests, it can
happen that single image block requests can get the result that is part
of loading lists of blocks. Since the response type can be empty here,
the respective image block loader functions need to be aware of this.