Admin: don't show new project/stack, stack group and mirror inline fields
So far, on the stack change page, we showed already by default a row of
input controls for new project/stack, stack group and stack mirror
objects. Because an error is shown when the stack object is saved and
the new inline object fields are left empty, no more empty rows are
shown. New objects can easily be added by clicking on the respective
link in the inline.